ecommerce/requirement.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.38.0
colorama==0.4.6
distlib==0.3.8
Django==5.0.1
django-active-link==0.1.8
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-shopping-cart==0.1
filelock==3.13.1
humanize==4.9.0
idna==3.6
include==0.2.2
kombu==5.3.5
MouseInfo==0.1.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.4
requests==2.31.0
six==1.16.0
slug==2.0
source==1.2.0
sqlparse==0.4.4
stripe==8.0.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13